Added wxLuaProcess to handle OnTerminate event to skip object deletion.
Using wxLuaProcess instead of wxProcess fixes a crash on OSX when the
application is exited after Detach call is made to the wxProcess object.

This happens because OnTerminate event supplied by wxwidgets will delete
the object, but the __gc handler in wxlua will try to do the same (the
deletion in wxwidgets is not triggered when END_PROCESS event is
handled, and when Detach is called, this doesn't happen).

If you want to replace wxProcess with wxLuaProcess, use the following:
  if wx.wxLuaProcess then wx.wxProcess = wx.wxLuaProcess end