Updated handling of integer values to better work with Lua 5.3+.
Previously, `math.type(wx.wxDisplaySize())` would return "float" under
Lua 5.3, but with this change, it should return "integer" as expected.

Some of the earlier work on this was done in 463112e5, but those changes
didn't include code in "override" files, which this patch addresses.

There are still instances where "number" is used (where wxIntPtr,
wxUIntPtr, and similar pointer types are used).